Update contacts.html
i have added contacts.html which works on form to send email to the library, it has a few features like name and email are required in order to send email. on CSS if you want to add style on head you can tell so that we remove contacts.css and the linking, also applies to js guys if want to add java code straight to contacts.html without linking.